Refactor TerminalClient to use pydantic types
