+++ tests/test_paystack.py
@@ -47,8 +47,8 @@ from pypaystack2.sub_clients.async_clients.settlements import AsyncSettlementCli
 from pypaystack2.sub_clients.async_clients.splits import AsyncTransactionSplitClient
 from pypaystack2.sub_clients.async_clients.subaccounts import AsyncSubAccountClient
 from pypaystack2.sub_clients.async_clients.subscriptions import AsyncSubscriptionClient
+from pypaystack2.sub_clients.async_clients.terminals import AsyncTerminalClient
 from pypaystack2.sub_clients.sync_clients.products import AsyncProductClient
-from pypaystack2.sub_clients.terminals import AsyncTerminalClient
 from pypaystack2.sub_clients.transactions import AsyncTransactionClient
 from pypaystack2.sub_clients.transfer_recipients import (
     TransferRecipientClient,
